Extract Breakpoint
