+++ modules/core/src/main/scala/lucuma/odb/api/model/SequenceModel.scala
@@ -3,7 +3,6 @@
 
 package lucuma.odb.api.model
 
-import lucuma.core.util.Enumerated
 import lucuma.odb.api.model.StepConfig.CreateStepConfig
 import lucuma.odb.api.model.syntax.inputvalidator._
 import cats.Eq
@@ -16,38 +15,6 @@ import monocle.{Iso, Lens}
 import monocle.macros.Lenses
 
 object SequenceModel {
-
-  sealed trait Breakpoint extends Product with Serializable {
-
-    def enabled: Boolean =
-      this match {
-        case Breakpoint.Enabled  => true
-        case Breakpoint.Disabled => false
-      }
-
-  }
-
-  object Breakpoint {
-
-    case object Enabled  extends Breakpoint
-    case object Disabled extends Breakpoint
-
-    val enabled: Breakpoint =
-      Enabled
-
-    val disabled: Breakpoint =
-      Disabled
-
-    val fromBoolean: Iso[Boolean, Breakpoint] =
-      Iso[Boolean, Breakpoint](b => if (b) Enabled else Disabled)(_.enabled)
-
-    implicit val EnumeratedBreakpoint: Enumerated[Breakpoint] =
-      Enumerated.of(enabled, disabled)
-
-    implicit val DecoderBreakpoint: Decoder[Breakpoint] =
-      deriveDecoder[Breakpoint]
-  }
-
 
   @Lenses final case class BreakpointStep[A](
     breakpoint: Breakpoint,
